feat(tree): finish tree slots: #icon and #content
Now we can use icon or content slots to pass svg,change to arrow or anything in checkbox to
check.But the is a question,the line's height left in tree is by calculating with current node's
chilren length.Sometimes it will be better? when we have three or more layer children.But if we
recursive traversal all children.the height calc is too large,we will see that the line is too
long,and out of the tree.That's ugly.